+++ packages/ts-newskit/src/components/slices/articleList/styles.ts
@@ -1,6 +1,0 @@
-import { styled, getSizingCssFromTheme, Block } from 'newskit';
-
-export const ContainerInline = styled(Block)`
-  display: inline-block;
-  ${getSizingCssFromTheme('height', 'sizing020')};
-`;